--- src/main/java/com/flow/flow/config/BaseException.java
@@ -1,13 +1,0 @@
-package com.flow.flow.config;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class BaseException extends Exception {
-    private BaseResponseStatus status;
-}
-
