+++ src/main/java/com/flow/model/GetCalendarRes.java
@@ -1,19 +1,14 @@
-package com.flow.flow.model;
+package com.flow.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.sql.Time;
-
 @Getter
 @Setter
 @AllArgsConstructor
-public class Transaction {
+public class GetCalendarRes {
+    private String date;
     private int isExp;
-    private String info;
     private int amount;
-    private String time;
-    private String memo;
 }
-
