+++ src/main/java/com/flow/model/GetFormReq.java
@@ -1,4 +1,4 @@
-package com.flow.flow.model;
+package com.flow.model;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
