+++ src/main/java/com/flow/model/GetTotalAmount.java
@@ -1,4 +1,4 @@
-package com.flow.flow.model;
+package com.flow.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,8 +7,8 @@ import lombok.Setter;
 @Getter
 @Setter
 @AllArgsConstructor
-public class GetCalendarRes {
-    private String date;
+public class GetTotalAmount {
+    private int total;
     private int isExp;
-    private int amount;
 }
+
