+++ src/main/java/com/flow/setting/SettingProvider.java
@@ -1,9 +1,9 @@
 
-package com.flow.flow.setting;
+package com.flow.setting;
 
-import com.flow.flow.config.BaseException;
-import com.flow.flow.model.Form;
-import com.flow.flow.model.GetFormRes;
+import com.flow.config.BaseException;
+import com.flow.model.Form;
+import com.flow.model.GetFormRes;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-import static com.flow.flow.config.BaseResponseStatus.DATABASE_ERROR;
+import static com.flow.config.BaseResponseStatus.DATABASE_ERROR;
 
 @Service
 public class SettingProvider {
